universal5420: hide EGL_KHR_surfaceless_context in opengl ES
Starting with Firefox 63, they used something in opengl EGL called EGL_KHR_surfaceless_context to render webpages.  For whatever reason in 16.0, the Mali GPU doesn't support this or work properly when this attribute is used so Firefox shows just white or blank pages.

There is a workaround for 16.0, but 2 ways to implement it.  The workaround is to hide the EGL_KHR_surfaceless_context attribute from Firefox either by
1) hex editing the binary blob file so that EGL_KHR_surfaceless_context is changed to, for example, EGL_160_surfaceless_context
2) go into the opengl code and tell the OS that the Mali GPU doesn't support EGL_KHR_surfaceless_context by commenting out that feature.

Change-Id: If14eb9ef04db666dcc42e2aa2a44ceb584f10e63