Revert "universal5420: kang mali blob from A800S ROM"
The blob, or the driver, has problems and reboots device due to OOM
when for example trying to play some games.  Until we figure out how
to fix this we revert to the old, stable, blob and driver.

this revert commit d0055c84ff6ac59f977b520cad64cf0ff6e53d91

Change-Id: I33bb56d71b07d93329533008b61ec7b26aec330c